+++ packages/frontend/src/components/ProjectConnection/DbtForms/BitBucketForm.tsx
@@ -7,7 +7,7 @@ import {
     isGitRepository,
     startWithSlash,
 } from '../../../utils/fieldValidators';
-import { useProjectFormContext } from '../ProjectFormProvider';
+import { useProjectFormContext } from '../ProjectFormProvider/useProjectFormContext';
 
 const BitBucketForm: FC<{ disabled: boolean }> = ({ disabled }) => {
     const { savedProject } = useProjectFormContext();
