+++ packages/frontend/src/components/ProjectConnection/DbtForms/GithubForm.tsx
@@ -8,7 +8,7 @@ import {
     isValidGithubToken,
     startWithSlash,
 } from '../../../utils/fieldValidators';
-import { useProjectFormContext } from '../ProjectFormProvider';
+import { useProjectFormContext } from '../ProjectFormProvider/useProjectFormContext';
 
 const GithubForm: FC<{ disabled: boolean }> = ({ disabled }) => {
     const { savedProject } = useProjectFormContext();
