+++ packages/frontend/src/components/ProjectConnection/ProjectFormProvider/index.tsx
@@ -1,0 +1,17 @@
+import { Project } from '@lightdash/common';
+import { createContext, FC } from 'react';
+
+export type ProjectFormContext = {
+    savedProject?: Project;
+};
+
+export const Context = createContext<ProjectFormContext | undefined>(undefined);
+
+export const ProjectFormProvider: FC<ProjectFormContext> = ({
+    savedProject,
+    children,
+}) => {
+    return (
+        <Context.Provider value={{ savedProject }}>{children}</Context.Provider>
+    );
+};
