+++ packages/frontend/src/components/ProjectConnection/WarehouseForms/SnowflakeForm.tsx
@@ -17,7 +17,7 @@ import {
 } from '../../../utils/fieldValidators';
 import FormSection from '../../ReactHookForm/FormSection';
 import FormCollapseButton from '../FormCollapseButton';
-import { useProjectFormContext } from '../ProjectFormProvider';
+import { useProjectFormContext } from '../ProjectFormProvider/useProjectFormContext';
 import StartOfWeekSelect from './Inputs/StartOfWeekSelect';
 
 export const SnowflakeSchemaInput: FC<{
